Fix reading upper 8 bits in spi_transaction
Previously, the uppermost byte of the three is read, but when shifting
it into the spi_transaction result value, it would overflow the 16 bits
used for this expression (due to integer promotion). This meant that
this byte would effectively always read as 0, which might not be a
problem in how spi_transaction was used so far.