Support hex files that are not at the end of flash
The previous code would handle an "end" record in the hex file by
terminating the current flash page, but not terminating the write
process. For bootloaders, which are aligned to the end of the flash,
this is not a problem, but for a normal sketch at the start of flash,
this causes a "no colon" error after the last page in the hex file is
written.

To handle this, an end record now causes readImagePage to return
nullptr, which signals write loop to stop writing.