Support automatically starting programming
This reads from the target's RESET pin to detect whether a target chip
or board is present, and automatically starts programming when a chip is
inserted. This requires a big pulldown resistor on the programmer side,
so the RESET pin reads low when no chip is inserted. This pulldown must
be big enough to not interfere with the target's pullup (so something like
1M).

Because a pullup is required, this feature is disable by default. It can
be enabled by changing the AUTOSTART define in the .ino file.