Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. Although the name of the repository does match the sketch name, when GitHub's popular Clone or download > Download ZIP feature is used to download the contents of a repository the branch/release/commit name is appended to the folder name, causing a mismatch.

When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "Standalone-Arduino-AVR-ISP-programmer.ino" needs to be inside a sketch folder named "Standalone-Arduino-AVR-ISP-programmer".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file Standalone-Arduino-AVR-ISP-programmer.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

Standalone-Arduino-AVR-ISP-programmer-master\Standalone-Arduino-AVR-ISP-programmer\Standalone-Arduino-AVR-ISP-programmer.ino:24:24: fatal error: optiLoader.h: No such file or directory